// src/types/api.ts
export interface JobRequest {
  title: string;
  department?: string;
  location: string;
  employmentType: "FULL_TIME" | "PART_TIME" | "CONTRACT" | "INTERNSHIP" | string;
  experience?: string;
  qualification?: string;
  description: string;
  responsibilities?: string;
  requirements?: string;
  displayOrder?: number;
  active?: boolean;
  applicationDeadline?: string;
}

export interface JobResponse {
  id: string;
  title: string;
  department: string;
  location: string;
  employmentType: "FULL_TIME" | "PART_TIME" | "CONTRACT" | "INTERNSHIP" | string;
  experience: string;
  qualification: string;
  description: string;
  responsibilities: string;
  requirements: string;
  displayOrder: number;
  active: boolean;
  applicationDeadline: string;
  createdAt: string;
  updatedAt: string;
}